More visualization stuff

+++ ros_packages/gcop_ctrl/cfg/cpp/gcop_ctrl/DMocInterfaceConfig.h
@@ -263,7 +263,7 @@ class DEFAULT
         boost::any val;
         (*i)->getValue(config, val);
 
-        if("Nit"==(*i)->name){Nit = boost::any_cast<double>(val);}
+        if("Nit"==(*i)->name){Nit = boost::any_cast<int>(val);}
         if("tf"==(*i)->name){tf = boost::any_cast<double>(val);}
         if("N"==(*i)->name){N = boost::any_cast<int>(val);}
         if("x0"==(*i)->name){x0 = boost::any_cast<double>(val);}
@@ -288,7 +288,7 @@ class DEFAULT
       }
     }
 
-    double Nit;
+    int Nit;
 double tf;
 int N;
 double x0;
@@ -320,7 +320,7 @@ double mu;
 
 
 //#line 259 "/opt/ros/fuerte/stacks/dynamic_reconfigure/src/dynamic_reconfigure/parameter_generator.py"
-      double Nit;
+      int Nit;
 //#line 259 "/opt/ros/fuerte/stacks/dynamic_reconfigure/src/dynamic_reconfigure/parameter_generator.py"
       double tf;
 //#line 259 "/opt/ros/fuerte/stacks/dynamic_reconfigure/src/dynamic_reconfigure/parameter_generator.py"
@@ -499,15 +499,15 @@ double mu;
     {
 DMocInterfaceConfig::GroupDescription<DMocInterfaceConfig::DEFAULT, DMocInterfaceConfig> Default("Default", "", 0, 0, true, &DMocInterfaceConfig::groups);
 //#line 259 "/opt/ros/fuerte/stacks/dynamic_reconfigure/src/dynamic_reconfigure/parameter_generator.py"
-      __min__.Nit = 0.0;
+      __min__.Nit = 0;
 //#line 259 "/opt/ros/fuerte/stacks/dynamic_reconfigure/src/dynamic_reconfigure/parameter_generator.py"
-      __max__.Nit = 30.0;
+      __max__.Nit = 30;
 //#line 259 "/opt/ros/fuerte/stacks/dynamic_reconfigure/src/dynamic_reconfigure/parameter_generator.py"
-      __default__.Nit = 10.0;
+      __default__.Nit = 10;
 //#line 259 "/opt/ros/fuerte/stacks/dynamic_reconfigure/src/dynamic_reconfigure/parameter_generator.py"
-      Default.abstract_parameters.push_back(DMocInterfaceConfig::AbstractParamDescriptionConstPtr(new DMocInterfaceConfig::ParamDescription<double>("Nit", "double", 0, "Number of iterations of dmoc", "", &DMocInterfaceConfig::Nit)));
+      Default.abstract_parameters.push_back(DMocInterfaceConfig::AbstractParamDescriptionConstPtr(new DMocInterfaceConfig::ParamDescription<int>("Nit", "int", 0, "Number of iterations of dmoc", "", &DMocInterfaceConfig::Nit)));
 //#line 259 "/opt/ros/fuerte/stacks/dynamic_reconfigure/src/dynamic_reconfigure/parameter_generator.py"
-      __param_descriptions__.push_back(DMocInterfaceConfig::AbstractParamDescriptionConstPtr(new DMocInterfaceConfig::ParamDescription<double>("Nit", "double", 0, "Number of iterations of dmoc", "", &DMocInterfaceConfig::Nit)));
+      __param_descriptions__.push_back(DMocInterfaceConfig::AbstractParamDescriptionConstPtr(new DMocInterfaceConfig::ParamDescription<int>("Nit", "int", 0, "Number of iterations of dmoc", "", &DMocInterfaceConfig::Nit)));
 //#line 259 "/opt/ros/fuerte/stacks/dynamic_reconfigure/src/dynamic_reconfigure/parameter_generator.py"
       __min__.tf = 0.0;
 //#line 259 "/opt/ros/fuerte/stacks/dynamic_reconfigure/src/dynamic_reconfigure/parameter_generator.py"
